Refactor: Point 구조체를 nested 타입이 아니라 전역 타입으로 선언
